backend/requirements.txt
appwrite
python-dotenv
Flask